refator: LottoRank의 금액 계산 파라미터 타입 수정
- wrapper 타입에서 primitive으로 변경

+++ src/main/java/lotto/domain/LottoRank.java
@@ -38,7 +38,7 @@ public enum LottoRank {
         return lottoRank.matchedCount == matchedCount;
     }
 
-    public long calculatePrize(Integer value) {
+    public long calculatePrize(long value) {
         return prize.calculateAmount(value);
     }
